pyproject.toml
[project]
name = "data-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.8.0",
    "dbt-duckdb>=1.8.0",
    "duckdb>=1.5.0",
    "ipykernel>=7.2.0",
    "pandas>=2.0.1",
    "pyarrow==17.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.56.0",
    "streamlit-echarts>=0.6.0",
]
